refactor(ui): extract sessions/useSessionsList hook from App.tsx
Moves the cursor-paged sessions list, its search-filter debounce, and the
load-more bookkeeping into a hook. setSessions stays exposed for the title
save/delete/title-suggest write paths that remain in App. Client draft
sessions and the fallback activity poll stay in App (routing-owned).
Pure move, no behavior change.
