+++ external/ui/src/ui/App.tsx
@@ -91,6 +91,7 @@ import { NavRail } from "./nav/NavRail";
 import { useShellLayout } from "./nav/useShellLayout";
 import { useLlmModelSelection } from "./chat/useLlmModelSelection";
 import { SessionsSidebar } from "./sessions/SessionsSidebar";
+import { useSessionsList } from "./sessions/useSessionsList";
 import { useConfirm } from "./components/useConfirm";
 import { useT } from "./i18n/I18nProvider";
 import type { SessionRow } from "./sessions/types";
@@ -197,10 +198,21 @@ export function App() {
   const [heroHomeGeneration, setHeroHomeGeneration] = useState(() =>
     Math.floor(Math.random() * HERO_ACCENT_VERBS.length),
   );
-  const [sessions, setSessions] = useState<SessionRow[]>([]);
-  const [sessionsCursor, setSessionsCursor] = useState<string | null>(null);
-  const sessionsCursorRef = useRef<string | null>(null);
-  const [sessionsError, setSessionsError] = useState<string | null>(null);
+  const headers = useMemo(
+    () => (sessionId ? { [HDR]: sessionId } : {}),
+    [sessionId],
+  );
+  const {
+    sessions,
+    setSessions,
+    sessionsError,
+    loadSessionsList,
+    sessionFilterDraft,
+    setSessionFilterDraft,
+    sessionFilterQ,
+    sessionsHasMore,
+    sessionsLoadingMore,
+  } = useSessionsList({ headers });
   const [items, setItems] = useState<TranscriptItem[]>([]);
   const [sessionLoading, setSessionLoading] = useState(false);
   const [sessionFadingOut, setSessionFadingOut] = useState(false);
@@ -489,12 +501,6 @@ export function App() {
   /** Set while the viewed session is a subagent's transcript (read-only, no composer). */
   const [subagentTranscript, setSubagentTranscript] =
     useState<SubagentTranscriptMeta | null>(null);
-  const [sessionFilterDraft, setSessionFilterDraft] = useState("");
-  const [sessionFilterQ, setSessionFilterQ] = useState("");
-  const [sessionsHasMore, setSessionsHasMore] = useState(false);
-  const [sessionsLoadingMore, setSessionsLoadingMore] = useState(false);
-  const sessionsHasMoreRef = useRef(false);
-  const sessionsLoadingMoreRef = useRef(false);
   const { viewportXL, railLabelsWide, toggleRailWidth } = useShellLayout();
   const [mode, setMode] = useState<string>("agent");
   const [describePreview, setDescribePreview] = useState<{
@@ -708,10 +714,6 @@ export function App() {
     );
   }
 
-  const headers = useMemo(
-    () => (sessionId ? { [HDR]: sessionId } : {}),
-    [sessionId],
-  );
 
   const {
     llmModelIds,
@@ -927,26 +929,6 @@ export function App() {
   }, [sessionId]);
 
   useEffect(() => {
-    const t = window.setTimeout(
-      () => setSessionFilterQ(sessionFilterDraft.trim()),
-      300,
-    );
-    return () => window.clearTimeout(t);
-  }, [sessionFilterDraft]);
-
-  useEffect(() => {
-    sessionsCursorRef.current = sessionsCursor;
-  }, [sessionsCursor]);
-
-  useEffect(() => {
-    sessionsHasMoreRef.current = sessionsHasMore;
-  }, [sessionsHasMore]);
-
-  useEffect(() => {
-    sessionsLoadingMoreRef.current = sessionsLoadingMore;
-  }, [sessionsLoadingMore]);
-
-  useEffect(() => {
     if (!sessionsOpen && !schedulerOpen) {
       return;
     }
@@ -970,68 +952,6 @@ export function App() {
     window.addEventListener("keydown", onKey);
     return () => window.removeEventListener("keydown", onKey);
   }, [sessionsOpen, schedulerOpen, schedulerEditor, closeSchedulerDrawer]);
-
-  const loadSessionsList = useCallback(
-    async (reset: boolean): Promise<SessionRow[] | null> => {
-      if (reset) {
-        sessionsCursorRef.current = null;
-        setSessionsCursor(null);
-      } else if (
-        !sessionsHasMoreRef.current ||
-        sessionsLoadingMoreRef.current
-      ) {
-        return null;
-      }
-      if (!reset) {
-        sessionsLoadingMoreRef.current = true;
-        setSessionsLoadingMore(true);
-      }
-      const ps = new URLSearchParams();
-      ps.set("limit", "30");
-      if (!reset) {
-        const cur = sessionsCursorRef.current;
-        if (cur) {
-          ps.set("cursor", cur);
-        }
-      }
-      if (sessionFilterQ) {
-        ps.set("q", sessionFilterQ);
-      }
-      ps.set("include_activity", "true");
-      const res = await fetchJSON<{
-        sessions: SessionRow[];
-        nextCursor?: string | null;
-        hasMore?: boolean;
-      }>(`/coddy/sessions?${ps.toString()}`, {
-        headers,
-      });
-      if (!reset) {
-        sessionsLoadingMoreRef.current = false;
-        setSessionsLoadingMore(false);
-      }
-      if (!res.ok || !res.data) {
-        setSessionsError(t("app.backendUnavailable", { status: res.status }));
-        return null;
-      }
-      setSessionsError(null);
-      const next = res.data.sessions || [];
-      setSessions((prev) => {
-        if (reset) {
-          return next;
-        }
-        const seen = new Set(prev.map((s) => s.id));
-        return [...prev, ...next.filter((s) => !seen.has(s.id))];
-      });
-      const nextCur = res.data.nextCursor ?? null;
-      setSessionsCursor(nextCur);
-      sessionsCursorRef.current = nextCur;
-      const hm = !!res.data.hasMore;
-      setSessionsHasMore(hm);
-      sessionsHasMoreRef.current = hm;
-      return next;
-    },
-    [sessionFilterQ, headers, t],
-  );
 
   useEffect(() => {
     void (async () => {
